converted mapped_addr to uint32_t

+++ Mapper.hpp
@@ -12,11 +12,11 @@ public:
 	Mapper(uint8_t prog_banks, uint8_t char_banks);
 	~Mapper();
 	
-	virtual uint8_t ppu_map_read(uint16_t addr, uint16_t *mapped_addr) = 0;
-	virtual uint8_t ppu_map_write(uint16_t addr, uint16_t *mapped_addr) = 0;
+	virtual uint8_t ppu_map_read(uint16_t addr, uint32_t *mapped_addr) = 0;
+	virtual uint8_t ppu_map_write(uint16_t addr, uint32_t *mapped_addr) = 0;
 
-	virtual uint8_t cpu_map_read(uint16_t addr, uint16_t *mapped_addr) = 0;
-	virtual uint8_t cpu_map_write(uint16_t addr, uint16_t *mapped_addr, uint8_t data) = 0;
+	virtual uint8_t cpu_map_read(uint16_t addr, uint32_t *mapped_addr) = 0;
+	virtual uint8_t cpu_map_write(uint16_t addr, uint32_t *mapped_addr, uint8_t data) = 0;
 
 };
